monitoring/requirements.txt
evidently==0.4.13
pandas==2.0.3
numpy==1.24.3
prometheus-client==0.19.0
redis==5.0.1
kafka-python==2.0.2

